[1509] `performance.Now()` returns double and client test (#1786)

+++ Html5/Performance.cs
@@ -15,7 +15,7 @@ namespace Bridge.Html5
         /// Returns a DOMHighResTimeStamp representing the amount of miliseconds elapsed since the start of the navigation, as give by PerformanceTiming.navigationStart to the call of the method.
         /// </summary>
         /// <returns></returns>
-        public virtual extern int Now();
+        public virtual extern double Now();
 
         /// <summary>
         /// Is a PerformanceTiming object containing latency-related performance information.
